refactor: 메시지 패키지 생성
- 메시지 관련 클래스를 한 곳에 모아 관리하기 위해 메시지 패키지를 생성합니다.
