+++ src/main/java/lotto/view/OutputView.java
@@ -2,6 +2,7 @@ package lotto.view;
 
 import java.util.List;
 import lotto.domain.Lotto;
+import lotto.message.PrintMessage;
 
 public class OutputView implements Output {
